pyproject.toml
[project]
name = "mlops-course-project"
version = "0.1.0"
description = "MLOps course project related to news aggregation"
dependencies = [
    "sentence-transformers>=4.1.0",
    "faiss-cpu>=1.10.0",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "nltk>=3.9.1",
    "pandarallel>=1.6.5",
    "clean-text>=0.6.0",
    "pymorphy3>=2.0.3",
    "optuna>=4.3.0",
    "hydra-core>=1.3.2",
    "snakemake>=9.5.1",
    "clearml>=2.0.0",
    "tritonclient[all]>=2.58.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ruff>=0.4.4",
    "pre-commit>=4.2.0",
    "mkdocs>=1.6.1",
    "skl2onnx>=1.19.1",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/mlops_course_project"]

[tool.ruff]
line-length = 88
target-version = "py312"
lint.select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "C90", # mccabe complexity
]
lint.fixable = ["ALL"]
lint.unfixable = []
extend-exclude = [".venv"]

[tool.ruff.lint.isort]
known-first-party = ["mlops_course_project"]

[tool.rye.scripts]
format = "ruff check . --fix"
lint = "ruff check ."
